backend/pyproject.toml
[tool.poetry]
name = "shiritori"
version = "0.1.0"
description = "Django REST API for Shiritori with Friends"
authors = ["Shiritori <coder@codertheory.dev>"]
readme = "README.md"
homepage = "shiritoriwithfriends.com"
repository = "https://github.com/codertheory/shiritori"
license = "MIT"

[tool.pycodestyle]
max-line-length = 120
exclude = [
    ".tox",
    ".git",
    "*/migrations/*",
    "*/static/CACHE/*",
    "docs",
    "node_modules",
    "venv",
    ".venv"
]

[tool.isort]
profile = "black"
line_length = 120
known_first_party = ["backend", "config"]
multi_line_output = 3
default_section = "THIRDPARTY"
skip = "venv/"
skip_glob = "**/migrations/*.py"
skip_gitignore = true
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true

[tool.mypy]
python_version = 3.11
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main"
]
exclude = [
    "migrations",
]

[tool.black]
line-length = 120

["tool.mypy-*.migrations.*"]
ignore_errors = true

[tool.coverage.run]
include = "backend/**"
omit = ["*migrations*", "*tests*"]
plugins = ["django_coverage_plugin"]

[tool.django-stubs]
django_settings_module = "config.settings.local"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=config.settings.test --reuse-db"
testpaths = [
    "tests",
    "integration",
]
python_files = [
    "test_*.py",
    "tests.py",
]

[tool.ruff]
exclude = [
    "migrations"
]
line-length = 120
# Assume Python 3.11.
target-version = "py311"

[tool.poetry.group.production]
optional = true

[tool.poetry.dependencies]
python = "^3.11"
pytz = "2022.7.1"  # https://github.com/stub42/pytz
python-slugify = "8.0.1"  # https://github.com/un33k/python-slugify
Pillow = "9.4.0"  # https://github.com/python-pillow/Pillow
argon2-cffi = "21.3.0"  # https://github.com/hynek/argon2_cffi
redis = "4.5.1"  # https://github.com/redis/redis-py
hiredis = "2.2.2"  # https://github.com/redis/hiredis-py
celery = "5.2.7"  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat = "2.5.0"  # https://github.com/celery/django-celery-beat
uvicorn = { version = "0.21.0", extras = ['standard'] }  # https://github.com/encode/uvicorn
nanoid = "2.0.0"
# Django
# ------------------------------------------------------------------------------
django = "4.1.7"  # pyup: < 4.1  # https://www.djangoproject.com/
django-environ = "0.10.0"  # https://github.com/joke2k/django-environ
django-model-utils = "4.3.1"  # https://github.com/jazzband/django-model-utils
django-redis = "5.2.0"  # https://github.com/jazzband/django-redis

# Django Channels
# ------------------------------------------------------------------------------
channels = { version = "4.0.0", extras = ['daphne'] }  #
channels-redis = "4.0.0"
daphne = "4.0.0"
Twisted = { version = "22.10.0", extras = ['tls', 'http2'] }  #

# Django REST Framework
djangorestframework = "3.14.0"  # https://github.com/encode/django-rest-framework
django-cors-headers = "3.14.0"  # https://github.com/adamchainz/django-cors-headers
djangorestframework-camel-case = "1.4.2" # https://github.com/vbabiy/djangorestframework-camel-case
# DRF-spectacular for api documentation
drf-spectacular = "0.26.0"  # https://github.com/tfranzel/drf-spectacular


[tool.poetry.group.dev.dependencies]
Werkzeug = { version = "^2.2.3", extras = ["watchdog"] } # https://github.com/pallets/werkzeug
ipdb = "^0.13.13" # https://github.com/gotcha/ipdb
psycopg2 = "^2.9.5"# https://github.com/psycopg/psycopg2
watchfiles = "^0.18.1" # https://github.com/samuelcolvin/watchfiles
# Testing
# ------------------------------------------------------------------------------
mypy = "^1.1.1"  # https://github.com/python/mypy
django-stubs = "^1.16.0" # https://github.com/typeddjango/django-stubs
djangorestframework-stubs = "^1.10.0"  # https://github.com/typeddjango/djangorestframework-stubs
pytest = "^7.2.2" # https://github.com/pytest-dev/pytest
pytest-sugar = "^0.9.6" # https://github.com/Frozenball/pytest-sugar
pytest-django = "^4.5.2" # https://github.com/pytest-dev/pytest-django
pytest-asyncio = "^0.20.3"  # https://github.com/pytest-dev/pytest-asyncio
pytest-mock = "^3.10.0" # https://github.com/pytest-dev/pytest-mock/
pytest-celery = "^0.0.0" # https://github.com/celery/pytest-celery
pytest-factoryboy = "^2.5.1" # https://github.com/pytest-dev/pytest-factoryboy
# Code quality
# ------------------------------------------------------------------------------
ruff = "^0.0.257"
coverage = "^7.2.2" # https://github.com/nedbat/coveragepy
black = "^23.1.0" # https://github.com/psf/black
pre-commit = "^3.2.0" # https://github.com/pre-commit/pre-commit
# Django
# ------------------------------------------------------------------------------
factory-boy = "^3.2.1" # https://github.com/FactoryBoy/factory_boy
django-extensions = "^3.2.1" # https://github.com/django-extensions/django-extensions
django-coverage-plugin = "^3.0.0"  # https://github.com/nedbat/django_coverage_plugin

[tool.poetry.group.production.dependencies]
gunicorn = "20.1.0"  # https://github.com/benoitc/gunicorn
psycopg2-binary = "2.9.5"  # https://github.com/psycopg/psycopg2
sentry-sdk = "1.17.0"  # https://github.com/getsentry/sentry-python
dj-database-url = "1.2.0" # https://github.com/jazzband/dj-database-url
